Split up callback.py
