+++ benchmarks/image_classification/mlx/train.py
@@ -25,14 +25,12 @@ from utils.logging import init_logging
 
 from pfl.aggregate.simulate import SimulatedBackend
 from pfl.aggregate.weighting import WeightByDatapoints, WeightByUser
-from pfl.callback import (
-    AggregateMetricsToDisk,
-    CentralEvaluationCallback,
-    ModelCheckpointingCallback,
-    StopwatchCallback,
-    TrackBestOverallMetrics,
-    WandbCallback,
-)
+from pfl.callback.aggregate_metrics_to_disk import AggregateMetricsToDisk
+from pfl.callback.central_evaluation import CentralEvaluationCallback
+from pfl.callback.checkpoint import ModelCheckpointingCallback
+from pfl.callback.stopwatch import StopwatchCallback
+from pfl.callback.track_best_overall_metrics import TrackBestOverallMetrics
+from pfl.callback.wandb import WandbCallback
 from pfl.hyperparam import NNEvalHyperParams, NNTrainHyperParams
 from pfl.internal.ops.mlx_ops import to_tensor
 from pfl.model.mlx import MLXModel
