+++ benchmarks/image_classification/tf/train.py
@@ -24,14 +24,12 @@ from utils.logging import init_logging
 from pfl.aggregate.simulate import SimulatedBackend
 from pfl.aggregate.weighting import WeightByDatapoints, WeightByUser
 from pfl.algorithm import FederatedAveraging, NNAlgorithmParams
-from pfl.callback import (
-    AggregateMetricsToDisk,
-    CentralEvaluationCallback,
-    ModelCheckpointingCallback,
-    StopwatchCallback,
-    TrackBestOverallMetrics,
-    WandbCallback,
-)
+from pfl.callback.aggregate_metrics_to_disk import AggregateMetricsToDisk
+from pfl.callback.central_evaluation import CentralEvaluationCallback
+from pfl.callback.checkpoint import ModelCheckpointingCallback
+from pfl.callback.stopwatch import StopwatchCallback
+from pfl.callback.track_best_overall_metrics import TrackBestOverallMetrics
+from pfl.callback.wandb import WandbCallback
 from pfl.hyperparam import NNEvalHyperParams, NNTrainHyperParams
 from pfl.model.tensorflow import TFModel
 from pfl.privacy import CentrallyAppliedPrivacyMechanism
