+++ benchmarks/llm/train.py
@@ -23,7 +23,10 @@ from utils.logging import init_logging
 
 from llm.argument_parsing import add_llm_arguments, parse_central_lr_scheduler, parse_peft_config
 from pfl.aggregate.simulate import SimulatedBackend
-from pfl.callback import AggregateMetricsToDisk, CentralEvaluationCallback, StopwatchCallback, WandbCallback
+from pfl.callback.aggregate_metrics_to_disk import AggregateMetricsToDisk
+from pfl.callback.central_evaluation import CentralEvaluationCallback
+from pfl.callback.stopwatch import StopwatchCallback
+from pfl.callback.wandb import WandbCallback
 from pfl.hyperparam import NNEvalHyperParams, NNTrainHyperParams
 from pfl.internal.ops import pytorch_ops
 from pfl.model.pytorch import PyTorchModel
