+++ benchmarks/lm/tf/train.py
@@ -25,15 +25,13 @@ from utils.logging import init_logging
 
 from pfl.aggregate.simulate import SimulatedBackend
 from pfl.algorithm import FederatedAveraging, NNAlgorithmParams
-from pfl.callback import (
-    AggregateMetricsToDisk,
-    CentralEvaluationCallback,
-    ModelCheckpointingCallback,
-    StopwatchCallback,
-    TensorBoardCallback,
-    TrackBestOverallMetrics,
-    WandbCallback,
-)
+from pfl.callback.aggregate_metrics_to_disk import AggregateMetricsToDisk
+from pfl.callback.central_evaluation import CentralEvaluationCallback
+from pfl.callback.checkpoint import ModelCheckpointingCallback
+from pfl.callback.stopwatch import StopwatchCallback
+from pfl.callback.tensorboard import TensorBoardCallback
+from pfl.callback.track_best_overall_metrics import TrackBestOverallMetrics
+from pfl.callback.wandb import WandbCallback
 from pfl.hyperparam import NNEvalHyperParams, NNTrainHyperParams
 from pfl.internal.platform import get_platform
 from pfl.model.tensorflow import TFModel
