+++ docs/source/reference/callback.rst
@@ -3,5 +3,38 @@
 Callbacks
 =========
 
-.. automodule:: pfl.callback
+.. automodule:: pfl.callback.aggregate_metrics_to_disk
+   :members:
+
+.. automodule:: pfl.callback.central_evaluation
+   :members:
+
+.. automodule:: pfl.callback.checkpoint
+   :members:
+
+.. automodule:: pfl.callback.convergence
+   :members:
+
+.. automodule:: pfl.callback.early_stopping
+   :members:
+
+.. automodule:: pfl.callback.profiler
+   :members:
+
+.. automodule:: pfl.callback.restore_training
+   :members:
+
+.. automodule:: pfl.callback.stopwatch
+   :members:
+
+.. automodule:: pfl.callback.tensorboard
+   :members:
+
+.. automodule:: pfl.callback.track_best_overall_metrics
+   :members:
+
+.. automodule:: pfl.callback.training_process
+   :members:
+
+.. automodule:: pfl.callback.wandb
    :members:
