+++ pfl/callback/__init__.py
@@ -1,0 +1,4 @@
+"""
+Callbacks to run during the training process.  See the base class
+`pfl.callback.training_process.TrainingProcessCallback`.
+"""
