+++ pfl/postprocessor/metrics.py
@@ -44,7 +44,7 @@ class SummaryMetrics(Postprocessor):
         There is no point in picking a smaller value ``<100`` for lower
         fidelity unless it is only for visualization purposes (the histograms
         generated by this postprocessor can be visualized with e.g.
-        :class:`~pfl.callback.TensorBoardCallback`).
+        :class:`~pfl.callback.tensorboard.TensorBoardCallback`).
 
         TODO: Add support for non-equally-spaced bins upon request.
           rdar://103595866 (histogram with non-equally-spaced bins)
