+++ publications/mdm/mdm_utils/utils/tools.py
@@ -1,7 +1,7 @@
 import os
 from typing import Tuple
 
-from pfl.callback import TrainingProcessCallback
+from pfl.callback.training_process import TrainingProcessCallback
 from pfl.internal.ops.selector import get_default_framework_module as get_ops
 from pfl.metrics import Metrics
 from pfl.model.base import StatefulModel
